# openapi.yaml
openapi: 3.1.0
info:
  title: SRM Credit Engine
  description: Motor de precificacao e registro auditavel de recebiveis adquiridos
    pelo FIDC.
  version: v1
servers:
- url: http://localhost:8080
  description: Generated server url
tags:
- name: Lotes de recebiveis
  description: "Recebimento, precificacao e consulta de lotes de recebiveis"
paths:
  /api/v1/lotes-recebiveis/{loteId}/recebiveis/{recebivelId}/liquidacao:
    put:
      tags:
      - Lotes de recebiveis
      summary: Liquida um recebivel precificado
      description: "Marca o recebivel como LIQUIDADO (paga o valorPresente ao cedente,\
        \ na moeda de pagamento). PUT em vez de POST deliberadamente: a operacao e'\
        \ idempotente - chamar novamente um recebivel ja liquidado (retry de rede,\
        \ duplo clique) retorna 200 com o mesmo resultado, sem gerar uma nova liquidacao\
        \ ou efeito colateral (ver SPEC.md)."
      operationId: liquidar
      parameters:
      - name: loteId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: recebivelId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: Recebivel liquidado (ou ja estava liquidado - idempotente)
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RecebivelResponse"
        "404":
          description: Lote ou recebivel nao encontrado
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RecebivelResponse"
        "422":
          description: Recebivel nao esta PRECIFICADO (nunca precificado ou rejeitado)
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RecebivelResponse"
  /api/v1/lotes-recebiveis:
    get:
      tags:
      - Lotes de recebiveis
      summary: "Lista os lotes de recebiveis, paginado"
      description: "Ordenacao por createdAt, dataReferencia ou status (formato: campo,asc|desc)."
      operationId: listar
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 20
      - name: sort
        in: query
        required: false
        schema:
          type: string
          default: "createdAt,desc"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PaginaResponseLoteRecebiveisResumoResponse"
    post:
      tags:
      - Lotes de recebiveis
      summary: Recebe e precifica um lote de recebiveis
      description: Calcula o desagio de cada recebivel do lote e registra o resultado
        de forma auditavel. Retorna 201 mesmo se algum recebivel for rejeitado individualmente
        (ver campo status de cada item).
      operationId: criar
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoteRecebiveisRequest"
        required: true
      responses:
        "201":
          description: Lote recebido e processado
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/LoteRecebiveisResponse"
  /api/v1/lotes-recebiveis/{id}:
    get:
      tags:
      - Lotes de recebiveis
      summary: "Busca o detalhe de um lote, com seus recebiveis precificados"
      operationId: buscarPorId
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/LoteRecebiveisResponse"
components:
  schemas:
    RecebivelResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        ativo:
          type: string
        valorBruto:
          type: string
          example: "15000.00"
        moeda:
          type: string
          enum:
          - BRL
          - USD
        dataVencimento:
          type: string
          format: date
        categoriaRisco:
          type: string
          enum:
          - AA
          - A
          - B
          - C
          - D
          - E
        status:
          type: string
          enum:
          - PENDENTE
          - PRECIFICADO
          - REJEITADO
          - LIQUIDADO
        valorPresente:
          type: string
          example: "14200.00"
        valorDesagio:
          type: string
          example: "800.00"
        taxaDescontoAplicada:
          type: string
          example: "0.146500"
        motivoRejeicao:
          type: string
        moedaPagamento:
          type: string
          description: Moeda em que o recebivel e' efetivamente pago; igual a `moeda`
            quando nao ha conversao cambial.
          enum:
          - BRL
          - USD
        cotacaoCambio:
          type: string
          description: Cotacao (BRL por 1 USD) usada na conversao; null quando moedaPagamento
            == moeda.
          example: "5.20"
        liquidadoEm:
          type: string
          format: date-time
          description: Data/hora em que o recebivel foi liquidado (pago ao cedente);
            null enquanto nao liquidado.
    LoteRecebiveisRequest:
      type: object
      properties:
        recebiveis:
          type: array
          items:
            $ref: "#/components/schemas/RecebivelRequest"
          minItems: 1
      required:
      - recebiveis
    RecebivelRequest:
      type: object
      properties:
        ativo:
          type: string
          minLength: 1
        valorBruto:
          type: string
          example: "15000.00"
        dataVencimento:
          type: string
          format: date
        categoriaRisco:
          type: string
          enum:
          - AA
          - A
          - B
          - C
          - D
          - E
        moedaPagamento:
          type: string
          description: "Moeda em que o recebivel e' efetivamente pago, se diferente\
            \ de BRL (cross-currency). Omitir quando o pagamento e' em BRL (mesma\
            \ moeda do ativo)."
          enum:
          - BRL
          - USD
      required:
      - ativo
      - categoriaRisco
      - dataVencimento
      - valorBruto
    LoteRecebiveisResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        dataReferencia:
          type: string
          format: date
        status:
          type: string
          enum:
          - RECEBIDO
          - PRECIFICADO
          - ERRO
        recebiveis:
          type: array
          items:
            $ref: "#/components/schemas/RecebivelResponse"
    LoteRecebiveisResumoResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        dataReferencia:
          type: string
          format: date
        status:
          type: string
          enum:
          - RECEBIDO
          - PRECIFICADO
          - ERRO
        createdAt:
          type: string
          format: date-time
    PaginaResponseLoteRecebiveisResumoResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/LoteRecebiveisResumoResponse"
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32